able to populate the shopping cart in the right hand side. layout works. Need a remove button
