+++ ecommerce-ui/src/rentals/VacationRental.jsx
@@ -1,21 +1,21 @@
 /* eslint-disable react/prop-types */
-import React, {useState} from 'react'
+import React from 'react'
 import VacationList from './VacationList'
-import ShoppingCart from "./ShoppingCart";
+// import ShoppingCart from "./ShoppingCart";
 // import {useState} from "@types/react";
 
-export default function VacationRental({bnbs}) {
+export default function VacationRental({bnbs, addToCart}) {
 
-  const [cart, setCart] = useState([])
-
-  function addToCart(card){
-
-    console.log('button clicked')
-    console.log(cart)
-
-    setCart([...cart, card])
-
-  }
+  // const [cart, setCart] = useState([])
+  //
+  // function addToCart(card){
+  //
+  //   console.log('button clicked')
+  //   console.log(cart)
+  //
+  //   setCart([...cart, card])
+  //
+  // }
 
   return (
     <div>
@@ -26,11 +26,12 @@ export default function VacationRental({bnbs}) {
 
       </div>
 
-      <div>
 
-        <ShoppingCart cart={cart}/>
+      {/*<div>*/}
 
-      </div>
+      {/*  <ShoppingCart cart={cart}/>*/}
+
+      {/*</div>*/}
 
 
     </div>
